Use urllib2 instead of urllib
urllib has trouble with certain proxy configurations, for example,
failing to retrieve the https pypi index when the proxy is set as http.
urllib2.urlopen does not have the same problem, and has the added benefit
of raising URLError when a server responds with a code > 400, which means
we can get rid of our custom error classes.